Get usage and limit information for Pushover applications

get_pushover_limits() retrieves the message usage and limit information for the given application.

get_pushover_limits(app = get_pushover_app())

Arguments

app
application token (see set_pushover_app())

Value

A list containing messaging usage for the given app. Fields include:

limit: Number of messages allowed per month

remaining: Number of remaining messages in current month

reset: Unix timestamp indicating when message count is reset

status}: request status (1= success) \itemrequest: unique request ID \item raw`: the raw httr::response object

Note

This information can alternatively be gotten by examining the headers in the response to previous API calls. Look for headers x-limit-app-limit, x-limit-app-remaining, and x-limit-app-reset. For example, if x stores the response from a pushover() call, httr::headers(x$raw) will return all of the headers included in the response.

Examples

if (FALSE) {
lims <- get_pushover_limits(app = "azGDORePK8gMaC0QOYAMyEEuzJnyUi")
}