Manage group subscriptions

These functions manage a user's membership in a Pushover delivery group

group_add_user() adds a user to a group. Optionally, a device can be specified on which that user will receive notifications

group_delete_user() removes a user from a group

group_disable_user() temporarily disables a user from receiving group notifications.

group_enable_user() re-enables a user to receive group notifications for a group

group_subscription(cmd, ...)

group_add_user(
  group,
  user,
  app = get_pushover_app(),
  device = NULL,
  memo = NULL
)

group_delete_user(group, user, app = get_pushover_app())

group_disable_user(group, user, app = get_pushover_app())

group_enable_user(group, user, app = get_pushover_app())

Arguments

cmd
The group subscription command to execute

...
Specific arguments for the given group subscription command

group
group key

user
user key

app
application token (see set_pushover_app())

device
(optional) device name to receive messages (defaults to all devices)

memo
(optional) memo about the user

Value

An invisible list containing the following fields:

status: request status (1 = success)

request: unique request ID

raw: the raw httr::response object

Examples

if (FALSE) {
group_add_user(
  group = "gznej3rKEVAvPUxu9vvNnqpmZpokzF",
  user = "uQiRzpo4DXghDmr9QzzfQu27cmVRsG",
  device = "phone"
)
}
if (FALSE) {
group_delete_user(
  group = "gznej3rKEVAvPUxu9vvNnqpmZpokzF",
  user = "uQiRzpo4DXghDmr9QzzfQu27cmVRsG"
)
}
if (FALSE) {
group_disable_user(
  group = "gznej3rKEVAvPUxu9vvNnqpmZpokzF",
  user = "uQiRzpo4DXghDmr9QzzfQu27cmVRsG"
)
}
if (FALSE) {
group_enable_user(
  group = "gznej3rKEVAvPUxu9vvNnqpmZpokzF",
  user = "uQiRzpo4DXghDmr9QzzfQu27cmVRsG"
)
}