Send a message using Pushover

pushover() sends a message (push notification) to a user or group. Messages can be given different priorities, play different sounds, or require acknowledgments. pushover_normal(), pushover_silent, pushover_quiet, pushover_high, and pushover_emergency functions send messages with those priorities.

pushover(
  message,
  title = NULL,
  priority = 0,
  user = get_pushover_user(),
  app = get_pushover_app(),
  device = NULL,
  sound = NULL,
  url = NULL,
  url_title = NULL,
  retry = 60,
  expire = 3600,
  callback = NULL,
  timestamp = NULL
)

pushover_silent(message, ...)

pushover_quiet(message, ...)

pushover_normal(message, ...)

pushover_high(message, ...)

pushover_emergency(message, ...)

Arguments

message
The message to be sent (max. 1024 characters)

title
(optional) The message's title

priority
Message priority (-2: silent, -1: quiet, 0: normal (default), 1: high, 2: emergency)

user
user/group key (see set_pushover_user())

app
application token (see set_pushover_app())

device
(optional) name of the device(s) to send message to. Defaults to all devices.

sound
(optional) name of the sound to play (see https://pushover.net/api#sounds)

url
(optional) supplementary URL to display with message

url_title
(optional) title to show for supplementary URL

retry
(optional) how often (in seconds) to repeat emergency priority messages (min: 30 seconds; default: 60 seconds)

expire
(optional) how long (in seconds) emergency priority messages will be retried (max: 86400 seconds; default: 3600 seconds)

callback
(optional) callback URL to be visited (HTTP POST) once an emergency priority message has been acknowledged (details)

timestamp
(optional) a Unix timestamp containing the date and time to display to the user instead of the time at which the message was received

...
Additional arguments to pass to pushover()

Value

an invisible list containing the following fields:

status: request status (1 = success)

request: unique request ID

raw: the raw httr::response object

receipt: a receipt ID (only for emergency priority messages)

errors: a list of error messages (only for unsuccessful requests)

Examples

if (FALSE) {
pushover(message = "Hola Mundo!")
}