Issue a command using the Pushover API

pushover_api() allows commands to be issued using the Pushover API. This is a generic function that is meant to be used by higher level functions. In most instances, more specific functions should be used (e.g., pushover()).

pushover_api(verb, url, visible = TRUE, ...)

Arguments

verb
The http method to use

url
The URL to visit

visible
Whether or not the result should be visible (default: TRUE)

...
Any additional parameters to be passed to httr::VERB

Value

a list containing the following fields and any other fields related to the specific API call:

status: request status (1 = success)

request: unique request ID

raw: the raw httr::response object

errors: a list of error messages (only for unsuccessful requests)

Examples

if (FALSE) {
pushover_api(
  verb = "GET",
  url = "https://api.pushover.net/1/sounds.json",
  query = list(token = "azGDORePK8gMaC0QOYAMyEEuzJnyUi")
)
}