Set, get, and unset the Pushover user/group key

set_pushover_user() sets the Pushover user or group key to be used in subsequent commands, get_pushover_user() gets the user or group key that is currently set, and unset_pushover_user() unsets the key. pushover_user.isset() indicates whether or not the user/group key has been set.

set_pushover_user(user = NULL, ask = interactive())

get_pushover_user(ask = interactive())

unset_pushover_user()

pushover_user.isset()

set_pushover_group(user = NULL, ask = interactive())

get_pushover_group(ask = interactive())

unset_pushover_group()

Arguments

user
The user or group key to be used. If none is provided, a prompt will request the key.

ask
Whether or not to ask for the key if none is provided. Note that this only works for interactive sessions.

Value

get_pushover_user() returns a string containing the current user or group key

pushover_user.isset() returns a logical value indicating whether or not the user/group is set.

Details

set_pushover_group(), get_pushover_group(), and unset_pushover_group() are aliases for these functions.

set_pushover_user() only sets the Pushover user or group for the current session. If a different value is specified in .Renviron, that value will be used in future sessions. Similarly, unset_pushover_user() will only unset the user or group for the current session.

User keys can be found within the settings of the Pushover app or by logging in to https://pushover.net. Group keys can be found after creating a delivery group in your account on https://pushover.net.

Examples

if (FALSE) {
set_pushover_user(user = "uQiRzpo4DXghDmr9QzzfQu27cmVRsG")
}