Validate Pushover values

is.valid_app() determines whether or not a given application token is valid or not according to Pushover's specifications. It does not determine whether or not the given token is associated with an application.

is.valid_device() only determines whether a device name is valid. Valid device names are strings up to 25 characters long and can include letters, numbers, _, and -. It does not determine whether that device is actually registered (see: is.registered_device().

is.valid_receipt() determines whether or not a given message receipt is valid or not according to Pushover's specifications. It does not determine whether or not the given receipt actually exists. Receipts are 30-character strings containing letters and numbers ([A-Za-z0-9]).

is.valid_user() and is.valid_group() determine whether or not a given user/group key is valid. They do not determine whether or not the key is registered with Pushover.

is.valid_sound() checks whether or not a given sound is supported by Pushover. See https://pushover.net/api#sounds for a complete list.

is.valid_app(token)

is.valid_device(device)

is.valid_receipt(receipt)

is.valid_user(user)

is.valid_group(group)

is.valid_sound(sound)

Arguments

token
A application token (e.g., "azGDORePK8gMaC0QOYAMyEEuzJnyUi")

device
one or more device names (e.g., "phone", c("phone", "tablet"))

receipt
A message receipt (e.g., "KAWXTswy4cekx6vZbHBKbCKk1c1fdf")

user
user/group key to verify (e.g., "uQiRzpo4DXghDmr9QzzfQu27cmVRsG")

group
user/group key to verify (e.g., "gznej3rKEVAvPUxu9vvNnqpmZpokzF")

sound
a string containing a sound name

Value

A logical value indicating whether or not the application token is valid

is.valid_device() returns logical a value indicating whether or not the corresponding device name is valid

is.valid_receipt() returns a boolean value for each given message receipt ID indicating whether or not that receipt ID is valid.

A logical value indicating whether or not the given user/group key is valid.

is.valid_sound() returns a logical value indicating whether or not the given sound is supported by Pushover.

Note

To acquire an application token, create an app at https://pushover.net/apps

See also

Examples

is.valid_app(token = "azGDORePK8gMaC0QOYAMyEEuzJnyUi")

#> [1] TRUE

is.valid_device("my_phone")

#> [1] TRUE

if (FALSE) {
is.valid_receipt(receipt = "KAWXTswy4cekx6vZbHBKbCKk1c1fdf")
}

if (FALSE) {
is.valid_user("uQiRzpo4DXghDmr9QzzfQu27cmVRsG")
}
is.valid_sound("cosmic")

#> [1] TRUE